+++ lib/yahaserver/yahaserver.h
@@ -12,17 +12,17 @@
 #pragma once
 
 #include <Arduino.h>
-#include <imessagebroker.h>
-#include <idevice.h>
-#include <wlan.h>
-#include <brokerproxy.h>
-#include <mqttserver.h>
-#include <eepromaccess.h>
-#include <runtime.h>
+#include "imessagebroker.h"
+#include "idevice.h"
+#include "wlan.h"
+#include "brokerproxy.h"
+#include "mqttserver.h"
+#include "eepromaccess.h"
+#include "runtime.h"
 
 class YahaServer : public IMessageBroker {
 public:
-    YahaServer() : _isBatteryMode(true), _isPowerOn(false) {
+    YahaServer() : _isBatteryMode(false), _isPowerOn(false) {
         MQTTServer::registerOnUpdateFunction(updateConfig);
         addDevice(&wlan);
         addDevice(&brokerProxy);
@@ -30,9 +30,9 @@ public:
 
     /**
      * Setup everything
-     * @param stationSSID name of the station set up, if WLAN connection is not possible
+     * @param APSSID name of the access point set up, if WLAN connection is not possible
      */
-    void setup(const String stationSSID = "YahaESP8266");
+    void setup(const String APSSID = "YahaESP8266");
 
     /**
      * Close down and go to sleep
@@ -44,9 +44,15 @@ public:
      */
     void loop();
 
-    void addDevice(IDevice* device) {
+    /**
+     * Adds a device
+     * @param device pointer to a device object
+     * @param priority setup priority (0, setup called after yahaserver setup, 1 => setup called before)
+     */
+    void addDevice(IDevice* device, uint8_t priority = 0) {
         device->setMessageBroker(this);
         _devices.push_back(device);
+        _priority.push_back(priority);
     }
 
     /**
@@ -71,10 +77,17 @@ public:
 
 private:
 
+    /**
+     * Setup all devices of a priority
+     * @param priority of the devices to set up
+     */
+    void setupDevices(uint8_t priority);
+
     static void setDeviceConfigFromJSON(jsonObject_t& config);
 
     static const uint16_t EEPROM_START_ADDR = 0;
     static std::vector<IDevice*> _devices;
+    static std::vector<uint8_t> _priority;
 
     bool _isBatteryMode;
     bool _isPowerOn;