+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # change log for yaha station
 
+## 0.3.0 2021-05-03 update
+
+- Added functionality for a rain sensor
+- Changed the password for the initial access point to be longer than 8 characters
+- Changed the used yaha-broker interface to version 1.0 (from version 0.0)
+- Now activates the AP, on power-on, if the WLAN connection is not successful
+
 ## 0.2.0 2020-05-31 update
 
 - Added functionality for irrigation